refactor: 重构架构相关代码到架构代码中
