+++ kernel/src/arch/loongarch64/loongarch64.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #define ARCH_MAX_IRQ_NUM 1024
+#define SCHED_HZ 250
 
 #include <libs/klibc.h>
 #include <arch/loongarch64/csr.h>
@@ -41,3 +42,5 @@ void write_barrier(void);
 
 void arch_enable_user_access();
 void arch_disable_user_access();
+bool arch_memory_region_usable(uint64_t addr, uint64_t len);
+uintptr_t arch_get_return_address(uint32_t level);
