+++ kernel/src/arch/x86_64/x86_64.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #define ARCH_MAX_IRQ_NUM 256
+#define SCHED_HZ 1000
 
 #include "arch/x86_64/asm.h"
 #include "arch/x86_64/core/normal.h"
@@ -46,3 +47,5 @@ void write_barrier(void);
 
 void arch_enable_user_access();
 void arch_disable_user_access();
+bool arch_memory_region_usable(uint64_t addr, uint64_t len);
+uintptr_t arch_get_return_address(uint32_t level);
