+++ kernel/src/task/signal.c
@@ -1,4 +1,5 @@
 #include <arch/arch.h>
+#include <arch/signal.h>
 #include <task/ptrace.h>
 #include <task/signal.h>
 
@@ -9,59 +10,14 @@
 
 #include <init/callbacks.h>
 
-#if defined(__x86_64__)
-#include <arch/x86_64/syscall/nr.h>
-#elif defined(__aarch64__)
-#include <arch/aarch64/syscall/nr.h>
-#elif defined(__riscv__)
-#include <arch/riscv64/syscall/nr.h>
-#elif defined(__loongarch64__)
-#include <arch/loongarch64/syscall/nr.h>
-#endif
-
 #define SIGNAL_MIN_SIGSET_SIZE sizeof(uint32_t)
 #define SIGNAL_MAX_SIGSET_SIZE sizeof(sigset_t)
 #define SIGNAL_MAX_MASKED_SIG ((int)(sizeof(sigset_t) * 8 - 1))
-
-#if defined(__x86_64__) || defined(__aarch64__)
-typedef struct signal_kernel_sigaction {
-    sighandler_t handler;
-    unsigned long flags;
-    void (*restorer)(void);
-    sigset_t mask;
-} signal_kernel_sigaction_t;
-#elif defined(__riscv__) || defined(__loongarch64__)
-typedef struct signal_kernel_sigaction {
-    sighandler_t handler;
-    unsigned long flags;
-    sigset_t mask;
-    void *unused;
-} signal_kernel_sigaction_t;
-#endif
 
 #define ERESTARTSYS 512
 #define ERESTARTNOINTR 513
 #define ERESTARTNOHAND 514
 #define ERESTART_RESTARTBLOCK 516
-
-#if defined(__x86_64__)
-#define SIGNAL_X64_SYSCALL_INS_LEN 2
-#define SIGNAL_X64_UC_SIGCONTEXT_SS 0x2
-#define SIGNAL_X64_UC_STRICT_RESTORE_SS 0x4
-#define SIGNAL_X64_TRAPNO_PAGE_FAULT 14
-#define SIGNAL_X64_RT_SIGRETURN_TRAMPOLINE_SIZE 9U
-#endif
-
-#if defined(__aarch64__)
-#define SIGNAL_AARCH64_RT_SIGRETURN_TRAMPOLINE_SIZE 8U
-#endif
-
-#if defined(__riscv__) || defined(__loongarch64__)
-#define SIGNAL_RISCV64_RT_SIGRETURN_TRAMPOLINE_SIZE 8U
-#define SIGNAL_RISCV64_SYSCALL_INS_LEN 4U
-#define SIGNAL_RISCV64_USER_SSTATUS ((1UL << 18) | (1UL << 13) | (1UL << 5))
-#define SIGNAL_RISCV64_FRAME_FROM_SYSCALL 1U
-#endif
 
 signal_internal_t signal_internal_decisions[MAXSIG] = {0};
 
@@ -189,21 +145,8 @@ static inline uint64_t signal_current_user_sp(task_t *task) {
     if (!task)
         return 0;
 
-#if defined(__x86_64__)
     struct pt_regs *regs = (struct pt_regs *)task->syscall_stack - 1;
-    return regs->rsp;
-#elif defined(__aarch64__)
-    struct pt_regs *regs = (struct pt_regs *)task->syscall_stack - 1;
-    return regs->sp_el0;
-#elif defined(__riscv__)
-    struct pt_regs *regs = (struct pt_regs *)task->syscall_stack - 1;
-    return regs->sp;
-#elif defined(__loongarch64__)
-    struct pt_regs *regs = (struct pt_regs *)task->syscall_stack - 1;
-    return regs->usp;
-#else
-    return 0;
-#endif
+    return signal_arch_user_sp_from_regs(regs);
 }
 
 static inline bool signal_is_blocked(sigset_t blocked, int sig) {
@@ -349,47 +292,6 @@ signal_has_deliverable_outside_set_locked(task_t *task, sigset_t wait_set) {
     return false;
 }
 
-static inline bool signal_arch_user_context(struct pt_regs *regs) {
-    if (!regs) {
-        return false;
-    }
-
-#if defined(__x86_64__)
-    return (regs->cs & 0x3) == 0x3;
-#elif defined(__aarch64__)
-    return (regs->cpsr & 0xF) == 0;
-#elif defined(__riscv__)
-    return (regs->sstatus & (1ULL << 8)) == 0;
-#elif defined(__loongarch64__)
-    return (regs->csr_prmd & 0x3) == 0;
-#else
-    return true;
-#endif
-}
-
-#if defined(__x86_64__)
-static const uint8_t signal_x64_rt_sigreturn_trampoline
-    [SIGNAL_X64_RT_SIGRETURN_TRAMPOLINE_SIZE] = {
-        0xB8, SYS_RT_SIGRETURN, 0x00, 0x00, 0x00, 0x0F, 0x05, 0x0F, 0x0B,
-};
-#endif
-
-#if defined(__aarch64__)
-static const uint8_t signal_aarch64_rt_sigreturn_trampoline
-    [SIGNAL_AARCH64_RT_SIGRETURN_TRAMPOLINE_SIZE] = {
-        0x68, 0x11, 0x80, 0xD2, // mov x8, #139
-        0x01, 0x00, 0x00, 0xD4, // svc 0
-};
-#endif
-
-#if defined(__riscv__) || defined(__loongarch64__)
-static const uint8_t signal_riscv64_rt_sigreturn_trampoline
-    [SIGNAL_RISCV64_RT_SIGRETURN_TRAMPOLINE_SIZE] = {
-        0x93, 0x08, 0xB0, 0x08, // li a7, 139
-        0x73, 0x00, 0x00, 0x00, // ecall
-};
-#endif
-
 static bool signal_sync_user_instruction_memory(uint64_t user_addr,
                                                 size_t code_size) {
     if (!user_addr || code_size == 0)
@@ -414,24 +316,6 @@ static bool signal_sync_user_instruction_memory(uint64_t user_addr,
     }
 
     return true;
-}
-
-#if defined(__x86_64__) || defined(__aarch64__) || defined(__riscv__) ||       \
-    defined(__loongarch64__)
-static const uint8_t *signal_arch_rt_sigreturn_trampoline(size_t *code_size) {
-    if (!code_size)
-        return NULL;
-
-#if defined(__x86_64__)
-    *code_size = sizeof(signal_x64_rt_sigreturn_trampoline);
-    return signal_x64_rt_sigreturn_trampoline;
-#elif defined(__aarch64__)
-    *code_size = sizeof(signal_aarch64_rt_sigreturn_trampoline);
-    return signal_aarch64_rt_sigreturn_trampoline;
-#elif defined(__riscv__) || defined(__loongarch64__)
-    *code_size = sizeof(signal_riscv64_rt_sigreturn_trampoline);
-    return signal_riscv64_rt_sigreturn_trampoline;
-#endif
 }
 
 static bool signal_ensure_user_trampoline(task_t *task) {
@@ -542,7 +426,6 @@ static bool signal_ensure_user_trampoline(task_t *task) {
 
     return true;
 }
-#endif
 
 static inline void signal_set_default(int sig, signal_internal_t action) {
     if (signal_sig_in_range(sig)) {
@@ -550,978 +433,7 @@ static inline void signal_set_default(int sig, signal_internal_t action) {
     }
 }
 
-#if defined(__x86_64__)
-enum {
-    X64_REG_R8 = 0,
-    X64_REG_R9,
-    X64_REG_R10,
-    X64_REG_R11,
-    X64_REG_R12,
-    X64_REG_R13,
-    X64_REG_R14,
-    X64_REG_R15,
-    X64_REG_RDI,
-    X64_REG_RSI,
-    X64_REG_RBP,
-    X64_REG_RBX,
-    X64_REG_RDX,
-    X64_REG_RAX,
-    X64_REG_RCX,
-    X64_REG_RSP,
-    X64_REG_RIP,
-    X64_REG_EFL,
-    X64_REG_CSGSFS,
-    X64_REG_ERR,
-    X64_REG_TRAPNO,
-    X64_REG_OLDMASK,
-    X64_REG_CR2,
-};
-
-static inline bool signal_x64_is_syscall_context(const struct pt_regs *regs) {
-    return ((regs->cs & 0x3) == 0x3) && regs->rip == regs->rcx &&
-           regs->rflags == regs->r11;
-}
-
-static inline void
-signal_x64_prepare_syscall_result(struct pt_regs *saved,
-                                  const sigaction_t *action) {
-    if (!signal_x64_is_syscall_context(saved)) {
-        return;
-    }
-
-    int64_t retval = (int64_t)saved->rax;
-    if (retval >= 0) {
-        return;
-    }
-
-    switch (-retval) {
-    case ERESTARTNOHAND:
-    case ERESTART_RESTARTBLOCK:
-        saved->rax = (uint64_t)-EINTR;
-        break;
-    case ERESTARTSYS:
-        if ((action->sa_flags & SA_RESTART) == 0) {
-            saved->rax = (uint64_t)-EINTR;
-            break;
-        }
-        __attribute__((fallthrough));
-    case ERESTARTNOINTR:
-        saved->rax = saved->orig_rax;
-        if (saved->rip >= SIGNAL_X64_SYSCALL_INS_LEN)
-            saved->rip -= SIGNAL_X64_SYSCALL_INS_LEN;
-        break;
-    default:
-        return;
-    }
-}
-
-typedef struct signal_x64_frame_layout {
-    uint64_t frame_rsp;
-    uint64_t ucontext_addr;
-    uint64_t siginfo_addr;
-    uint64_t fpstate_addr;
-    uint64_t fpstate_bytes;
-    uint64_t trampoline_addr;
-    uint64_t trampoline_bytes;
-} signal_x64_frame_layout_t;
-
-static inline uint64_t signal_x64_align_up(uint64_t value, uint64_t align) {
-    return (value + align - 1) & ~(align - 1);
-}
-
-static uint64_t signal_x64_saved_fpstate_bytes(void) {
-    uint64_t bytes = x64_fpu_state_size();
-
-    if (bytes < sizeof(fpu_context_t))
-        bytes = sizeof(fpu_context_t);
-
-    return bytes;
-}
-
-static uint64_t signal_x64_user_fpstate_bytes(void) {
-    uint64_t bytes = signal_x64_saved_fpstate_bytes();
-
-    if (x64_fpu_xsave_enabled())
-        bytes += X64_FP_XSTATE_MAGIC2_SIZE;
-
-    return bytes;
-}
-
-static void signal_x64_build_frame_layout(uint64_t user_rsp,
-                                          uint64_t fpstate_bytes,
-                                          uint64_t trampoline_bytes,
-                                          signal_x64_frame_layout_t *layout) {
-    if (!layout)
-        return;
-
-    memset(layout, 0, sizeof(*layout));
-
-    uint64_t redzone_top = user_rsp - 128;
-    uint64_t fixed_bytes = sizeof(void *) + sizeof(ucontext_t) +
-                           sizeof(siginfo_t) + trampoline_bytes;
-    uint64_t reserve = fixed_bytes + (X64_FPU_FRAME_ALIGN - 1) + fpstate_bytes;
-
-    if (redzone_top <= reserve)
-        return;
-
-    uint64_t frame_rsp = redzone_top - reserve;
-    frame_rsp =
-        (frame_rsp & ~(uint64_t)(X64_FPU_FRAME_ALIGN - 1)) + sizeof(void *);
-    if (frame_rsp > redzone_top - reserve)
-        frame_rsp -= X64_FPU_FRAME_ALIGN;
-
-    layout->frame_rsp = frame_rsp;
-    layout->ucontext_addr = frame_rsp + sizeof(void *);
-    layout->siginfo_addr = layout->ucontext_addr + sizeof(ucontext_t);
-    layout->fpstate_addr = signal_x64_align_up(
-        layout->siginfo_addr + sizeof(siginfo_t), X64_FPU_FRAME_ALIGN);
-    layout->fpstate_bytes = fpstate_bytes;
-    layout->trampoline_addr = layout->fpstate_addr + fpstate_bytes;
-    layout->trampoline_bytes = trampoline_bytes;
-}
-
-static bool signal_x64_copy_fpstate_to_user(task_t *task,
-                                            uint64_t user_fpstate_addr,
-                                            uint64_t fpstate_bytes) {
-    if (!task || !task->arch_context || !task->arch_context->fpu_ctx)
-        return false;
-
-    uint64_t saved_bytes = signal_x64_saved_fpstate_bytes();
-    void *image = malloc(fpstate_bytes);
-    if (!image)
-        return false;
-
-    memset(image, 0, fpstate_bytes);
-    memcpy(image, task->arch_context->fpu_ctx, saved_bytes);
-
-    fpu_context_t *fp = (fpu_context_t *)image;
-    if (x64_fpu_xsave_enabled()) {
-        uint64_t xfeatures = X64_XSTATE_X87 | X64_XSTATE_SSE;
-        if (saved_bytes > X64_XSAVE_HDR_OFFSET) {
-            const x64_xsave_header_t *hdr =
-                (const x64_xsave_header_t *)((const uint8_t *)image +
-                                             X64_XSAVE_HDR_OFFSET);
-            if (hdr->xstate_bv)
-                xfeatures = hdr->xstate_bv;
-        }
-
-        fp->sw_reserved.magic1 = X64_FP_XSTATE_MAGIC1;
-        fp->sw_reserved.extended_size = (uint32_t)fpstate_bytes;
-        fp->sw_reserved.xstate_bv = xfeatures;
-        fp->sw_reserved.xstate_size = (uint32_t)saved_bytes;
-        *(uint32_t *)((uint8_t *)image + fpstate_bytes -
-                      X64_FP_XSTATE_MAGIC2_SIZE) = X64_FP_XSTATE_MAGIC2;
-    } else {
-        memset(&fp->sw_reserved, 0, sizeof(fp->sw_reserved));
-    }
-
-    bool ok = !copy_to_user((void *)user_fpstate_addr, image, fpstate_bytes);
-    free(image);
-    return ok;
-}
-
-static uint64_t signal_x64_restore_fpstate_bytes(const ucontext_t *ucontext) {
-    if (!ucontext)
-        return sizeof(fpu_context_t);
-
-    if ((ucontext->uc_flags & X64_UC_FP_XSTATE) && x64_fpu_xsave_enabled())
-        return signal_x64_saved_fpstate_bytes();
-
-    return sizeof(fpu_context_t);
-}
-
-static inline void signal_x64_fill_ucontext(ucontext_t *ucontext,
-                                            const struct pt_regs *regs,
-                                            sigset_t blocked_mask,
-                                            const stack_t *altstack,
-                                            const fpu_context_t *fpu_ctx) {
-    memset(ucontext, 0, sizeof(*ucontext));
-
-    ucontext->uc_flags =
-        SIGNAL_X64_UC_SIGCONTEXT_SS | SIGNAL_X64_UC_STRICT_RESTORE_SS;
-    if (x64_fpu_xsave_enabled())
-        ucontext->uc_flags |= X64_UC_FP_XSTATE;
-    ucontext->uc_link = NULL;
-    if (altstack)
-        ucontext->uc_stack = *altstack;
-    if (fpu_ctx) {
-        memcpy(&ucontext->__fpregs_mem, fpu_ctx,
-               sizeof(ucontext->__fpregs_mem));
-    }
-    ucontext->uc_mcontext.fpregs = NULL;
-
-    ucontext->uc_mcontext.gregs[X64_REG_R8] = regs->r8;
-    ucontext->uc_mcontext.gregs[X64_REG_R9] = regs->r9;
-    ucontext->uc_mcontext.gregs[X64_REG_R10] = regs->r10;
-    ucontext->uc_mcontext.gregs[X64_REG_R11] = regs->r11;
-    ucontext->uc_mcontext.gregs[X64_REG_R12] = regs->r12;
-    ucontext->uc_mcontext.gregs[X64_REG_R13] = regs->r13;
-    ucontext->uc_mcontext.gregs[X64_REG_R14] = regs->r14;
-    ucontext->uc_mcontext.gregs[X64_REG_R15] = regs->r15;
-    ucontext->uc_mcontext.gregs[X64_REG_RDI] = regs->rdi;
-    ucontext->uc_mcontext.gregs[X64_REG_RSI] = regs->rsi;
-    ucontext->uc_mcontext.gregs[X64_REG_RBP] = regs->rbp;
-    ucontext->uc_mcontext.gregs[X64_REG_RBX] = regs->rbx;
-    ucontext->uc_mcontext.gregs[X64_REG_RDX] = regs->rdx;
-    ucontext->uc_mcontext.gregs[X64_REG_RAX] = regs->rax;
-    ucontext->uc_mcontext.gregs[X64_REG_RCX] = regs->rcx;
-    ucontext->uc_mcontext.gregs[X64_REG_RSP] = regs->rsp;
-    ucontext->uc_mcontext.gregs[X64_REG_RIP] = regs->rip;
-    ucontext->uc_mcontext.gregs[X64_REG_EFL] = regs->rflags;
-    ucontext->uc_mcontext.gregs[X64_REG_CSGSFS] =
-        (regs->cs & 0xFFFFULL) | ((uint64_t)0 << 16) | ((uint64_t)0 << 32) |
-        ((regs->ss & 0xFFFFULL) << 48);
-    ucontext->uc_mcontext.gregs[X64_REG_ERR] = regs->errcode;
-    ucontext->uc_mcontext.gregs[X64_REG_TRAPNO] = 0;
-    ucontext->uc_mcontext.gregs[X64_REG_OLDMASK] = 0;
-    ucontext->uc_mcontext.gregs[X64_REG_CR2] = 0;
-    memset(&ucontext->uc_sigmask, 0, sizeof(ucontext->uc_sigmask));
-    ucontext->uc_sigmask.__bits[0] = sigset_kernel_to_user(blocked_mask);
-}
-
-static inline void signal_x64_restore_ptregs(struct pt_regs *regs,
-                                             const ucontext_t *ucontext) {
-    regs->r8 = ucontext->uc_mcontext.gregs[X64_REG_R8];
-    regs->r9 = ucontext->uc_mcontext.gregs[X64_REG_R9];
-    regs->r10 = ucontext->uc_mcontext.gregs[X64_REG_R10];
-    regs->r11 = ucontext->uc_mcontext.gregs[X64_REG_R11];
-    regs->r12 = ucontext->uc_mcontext.gregs[X64_REG_R12];
-    regs->r13 = ucontext->uc_mcontext.gregs[X64_REG_R13];
-    regs->r14 = ucontext->uc_mcontext.gregs[X64_REG_R14];
-    regs->r15 = ucontext->uc_mcontext.gregs[X64_REG_R15];
-    regs->rdi = ucontext->uc_mcontext.gregs[X64_REG_RDI];
-    regs->rsi = ucontext->uc_mcontext.gregs[X64_REG_RSI];
-    regs->rbp = ucontext->uc_mcontext.gregs[X64_REG_RBP];
-    regs->rbx = ucontext->uc_mcontext.gregs[X64_REG_RBX];
-    regs->rdx = ucontext->uc_mcontext.gregs[X64_REG_RDX];
-    regs->rax = ucontext->uc_mcontext.gregs[X64_REG_RAX];
-    regs->rcx = ucontext->uc_mcontext.gregs[X64_REG_RCX];
-    regs->rsp = ucontext->uc_mcontext.gregs[X64_REG_RSP];
-    regs->rip = ucontext->uc_mcontext.gregs[X64_REG_RIP];
-    regs->rflags = ucontext->uc_mcontext.gregs[X64_REG_EFL];
-    uint64_t csgsfs = ucontext->uc_mcontext.gregs[X64_REG_CSGSFS];
-    regs->cs = csgsfs & 0xFFFFULL;
-    regs->ss = (csgsfs >> 48) & 0xFFFFULL;
-    regs->errcode = ucontext->uc_mcontext.gregs[X64_REG_ERR];
-    regs->func = 0;
-
-    if ((regs->cs & 0x3) != 0x3) {
-        regs->cs = SELECTOR_USER_CS;
-    }
-    if ((regs->ss & 0x3) != 0x3) {
-        regs->ss = SELECTOR_USER_DS;
-    }
-}
-
-static bool signal_x64_setup_frame(task_t *task, struct pt_regs *regs, int sig,
-                                   const sigaction_t *action,
-                                   const siginfo_t *info,
-                                   sigset_t restore_mask) {
-    struct pt_regs saved = *regs;
-    signal_x64_prepare_syscall_result(&saved, action);
-
-    stack_t frame_altstack;
-    signal_altstack_format_old(&frame_altstack, &task->signal->altstack,
-                               saved.rsp);
-
-    uint64_t fpstate_bytes = signal_x64_user_fpstate_bytes();
-    uint64_t trampoline_bytes = 0;
-    bool use_kernel_restorer =
-        !(action->sa_flags & SA_RESTORER) || !action->sa_restorer;
-    if (use_kernel_restorer && !signal_ensure_user_trampoline(task))
-        return false;
-
-    signal_x64_frame_layout_t layout;
-    signal_x64_build_frame_layout(saved.rsp, fpstate_bytes, trampoline_bytes,
-                                  &layout);
-    bool use_altstack =
-        (action->sa_flags & SA_ONSTACK) &&
-        signal_altstack_config_enabled(&task->signal->altstack) &&
-        !signal_altstack_contains_sp(&task->signal->altstack, saved.rsp);
-    if (use_altstack) {
-        uint64_t alt_top = signal_stack_base(&task->signal->altstack) +
-                           task->signal->altstack.ss_size;
-        signal_x64_build_frame_layout(alt_top, fpstate_bytes, trampoline_bytes,
-                                      &layout);
-        if (task->signal->altstack.ss_flags & SS_AUTODISARM)
-            signal_altstack_disable(&task->signal->altstack);
-    }
-    if (!layout.frame_rsp)
-        return false;
-
-    uint64_t ucontext_addr = layout.ucontext_addr;
-    uint64_t siginfo_addr = layout.siginfo_addr;
-
-    ucontext_t frame_ucontext;
-    x64_fpu_save(task->arch_context->fpu_ctx);
-    signal_x64_fill_ucontext(&frame_ucontext, &saved, restore_mask,
-                             &frame_altstack, task->arch_context->fpu_ctx);
-    if (sig == SIGSEGV && info) {
-        frame_ucontext.uc_mcontext.gregs[X64_REG_TRAPNO] =
-            SIGNAL_X64_TRAPNO_PAGE_FAULT;
-        frame_ucontext.uc_mcontext.gregs[X64_REG_CR2] =
-            (uint64_t)info->_sifields._sigfault._addr;
-    }
-    frame_ucontext.uc_mcontext.fpregs = (fpu_context_t *)layout.fpstate_addr;
-
-    void *frame_restorer = (void *)action->sa_restorer;
-    if (use_kernel_restorer)
-        frame_restorer = (void *)USER_SIGNAL_TRAMPOLINE_START;
-    if (!signal_x64_copy_fpstate_to_user(task, layout.fpstate_addr,
-                                         layout.fpstate_bytes)) {
-        return false;
-    }
-
-    if (copy_to_user((void *)siginfo_addr, info, sizeof(*info)) ||
-        copy_to_user((void *)ucontext_addr, &frame_ucontext,
-                     sizeof(frame_ucontext)) ||
-        copy_to_user((void *)layout.frame_rsp, &frame_restorer,
-                     sizeof(frame_restorer))) {
-        return false;
-    }
-
-    memset(regs, 0, sizeof(*regs));
-    regs->rip = (uint64_t)action->sa_handler;
-    regs->rdi = sig;
-    if (action->sa_flags & SA_SIGINFO) {
-        regs->rsi = siginfo_addr;
-        regs->rdx = ucontext_addr;
-    }
-    regs->cs = SELECTOR_USER_CS;
-    regs->ss = SELECTOR_USER_DS;
-    regs->rflags = saved.rflags | (1ULL << 9);
-    regs->rsp = layout.frame_rsp;
-    regs->rbp = layout.frame_rsp;
-
-    return true;
-}
-#endif
-
-#if defined(__aarch64__)
-typedef struct signal_aarch64_user_sigset {
-    uint64_t __bits[16];
-} signal_aarch64_user_sigset_t;
-
-#define SIGNAL_AARCH64_RESERVED_BYTES 4096U
-#define SIGNAL_AARCH64_FPSIMD_MAGIC 0x46508001U
-#define SIGNAL_AARCH64_SYSCALL_INS_LEN 4U
-
-typedef struct signal_aarch64_ctx_header {
-    uint32_t magic;
-    uint32_t size;
-} signal_aarch64_ctx_header_t;
-
-typedef struct signal_aarch64_fpsimd_context {
-    signal_aarch64_ctx_header_t head;
-    uint32_t fpsr;
-    uint32_t fpcr;
-    uint64_t vregs[32][2];
-} __attribute__((aligned(16))) signal_aarch64_fpsimd_context_t;
-
-typedef struct signal_aarch64_sigcontext {
-    uint64_t fault_address;
-    uint64_t regs[31];
-    uint64_t sp;
-    uint64_t pc;
-    uint64_t pstate;
-    uint8_t __reserved[SIGNAL_AARCH64_RESERVED_BYTES]
-        __attribute__((aligned(16)));
-} signal_aarch64_sigcontext_t;
-
-typedef struct signal_aarch64_ucontext {
-    uint64_t uc_flags;
-    struct signal_aarch64_ucontext *uc_link;
-    stack_t uc_stack;
-    signal_aarch64_user_sigset_t uc_sigmask;
-    signal_aarch64_sigcontext_t uc_mcontext;
-} signal_aarch64_ucontext_t;
-
-typedef struct signal_aarch64_frame {
-    siginfo_t info;
-    signal_aarch64_ucontext_t ucontext;
-} signal_aarch64_frame_t;
-
-typedef struct signal_aarch64_frame_record {
-    uint64_t fp;
-    uint64_t lr;
-} signal_aarch64_frame_record_t;
-
-_Static_assert(sizeof(signal_aarch64_fpsimd_context_t) == 0x210,
-               "AArch64 fpsimd_context must match the Linux ABI");
-_Static_assert(offsetof(signal_aarch64_sigcontext_t, __reserved) == 0x120,
-               "AArch64 sigcontext.__reserved offset must match the Linux ABI");
-_Static_assert(sizeof(signal_aarch64_user_sigset_t) == 128,
-               "AArch64 user sigset in ucontext must be 1024 bits");
-
-static inline uint64_t signal_aarch64_align_down(uint64_t value,
-                                                 uint64_t align) {
-    return value & ~(align - 1);
-}
-
-static inline void
-signal_aarch64_save_fpsimd(signal_aarch64_sigcontext_t *sigcontext,
-                           const fpu_context_t *fpu_ctx) {
-    memset(sigcontext->__reserved, 0, sizeof(sigcontext->__reserved));
-    if (!fpu_ctx)
-        return;
-
-    signal_aarch64_fpsimd_context_t *fpsimd =
-        (signal_aarch64_fpsimd_context_t *)sigcontext->__reserved;
-    fpsimd->head.magic = SIGNAL_AARCH64_FPSIMD_MAGIC;
-    fpsimd->head.size = sizeof(*fpsimd);
-    fpsimd->fpsr = (uint32_t)fpu_ctx->fpsr;
-    fpsimd->fpcr = (uint32_t)fpu_ctx->fpcr;
-    memcpy(fpsimd->vregs, fpu_ctx->q, sizeof(fpsimd->vregs));
-}
-
-static bool
-signal_aarch64_restore_fpsimd(const signal_aarch64_sigcontext_t *sigcontext,
-                              fpu_context_t *fpu_ctx) {
-    if (!fpu_ctx)
-        return true;
-
-    aarch64_fpu_state_init(fpu_ctx);
-
-    const uint8_t *cursor = sigcontext->__reserved;
-    size_t remaining = sizeof(sigcontext->__reserved);
-    bool fpsimd_seen = false;
-
-    while (remaining >= sizeof(signal_aarch64_ctx_header_t)) {
-        const signal_aarch64_ctx_header_t *head =
-            (const signal_aarch64_ctx_header_t *)cursor;
-
-        if (head->magic == 0 && head->size == 0)
-            return true;
-
-        if ((head->size & 0xFUL) != 0 ||
-            head->size < sizeof(signal_aarch64_ctx_header_t) ||
-            head->size > remaining) {
-            return false;
-        }
-
-        if (head->magic == SIGNAL_AARCH64_FPSIMD_MAGIC) {
-            if (fpsimd_seen ||
-                head->size != sizeof(signal_aarch64_fpsimd_context_t)) {
-                return false;
-            }
-
-            const signal_aarch64_fpsimd_context_t *fpsimd =
-                (const signal_aarch64_fpsimd_context_t *)cursor;
-            fpu_ctx->fpsr = fpsimd->fpsr;
-            fpu_ctx->fpcr = fpsimd->fpcr;
-            memcpy(fpu_ctx->q, fpsimd->vregs, sizeof(fpsimd->vregs));
-            fpsimd_seen = true;
-        }
-
-        cursor += head->size;
-        remaining -= head->size;
-    }
-
-    return false;
-}
-
-static inline void
-signal_aarch64_prepare_syscall_result(struct pt_regs *saved,
-                                      const sigaction_t *action) {
-    if (!saved || !action || saved->syscallno == NO_SYSCALL ||
-        saved->syscallno == SYS_RT_SIGRETURN)
-        return;
-
-    int64_t retval = (int64_t)saved->x0;
-    if (retval >= 0)
-        return;
-
-    switch (-retval) {
-    case ERESTARTNOHAND:
-    case ERESTART_RESTARTBLOCK:
-        saved->x0 = (uint64_t)-EINTR;
-        break;
-    case ERESTARTSYS:
-        if ((action->sa_flags & SA_RESTART) == 0) {
-            saved->x0 = (uint64_t)-EINTR;
-            break;
-        }
-        __attribute__((fallthrough));
-    case ERESTARTNOINTR:
-        saved->x0 = saved->origin_x0;
-        if (saved->pc >= SIGNAL_AARCH64_SYSCALL_INS_LEN)
-            saved->pc -= SIGNAL_AARCH64_SYSCALL_INS_LEN;
-        break;
-    default:
-        return;
-    }
-}
-
-static inline void
-signal_aarch64_fill_ucontext(signal_aarch64_ucontext_t *ucontext,
-                             const struct pt_regs *regs, sigset_t blocked_mask,
-                             const stack_t *altstack, const siginfo_t *info,
-                             const fpu_context_t *fpu_ctx) {
-    memset(ucontext, 0, sizeof(*ucontext));
-
-    if (altstack)
-        ucontext->uc_stack = *altstack;
-    ucontext->uc_sigmask.__bits[0] = sigset_kernel_to_user(blocked_mask);
-
-    if (info && info->si_signo == SIGSEGV)
-        ucontext->uc_mcontext.fault_address =
-            (uint64_t)info->_sifields._sigfault._addr;
-
-    ucontext->uc_mcontext.regs[0] = regs->x0;
-    ucontext->uc_mcontext.regs[1] = regs->x1;
-    ucontext->uc_mcontext.regs[2] = regs->x2;
-    ucontext->uc_mcontext.regs[3] = regs->x3;
-    ucontext->uc_mcontext.regs[4] = regs->x4;
-    ucontext->uc_mcontext.regs[5] = regs->x5;
-    ucontext->uc_mcontext.regs[6] = regs->x6;
-    ucontext->uc_mcontext.regs[7] = regs->x7;
-    ucontext->uc_mcontext.regs[8] = regs->x8;
-    ucontext->uc_mcontext.regs[9] = regs->x9;
-    ucontext->uc_mcontext.regs[10] = regs->x10;
-    ucontext->uc_mcontext.regs[11] = regs->x11;
-    ucontext->uc_mcontext.regs[12] = regs->x12;
-    ucontext->uc_mcontext.regs[13] = regs->x13;
-    ucontext->uc_mcontext.regs[14] = regs->x14;
-    ucontext->uc_mcontext.regs[15] = regs->x15;
-    ucontext->uc_mcontext.regs[16] = regs->x16;
-    ucontext->uc_mcontext.regs[17] = regs->x17;
-    ucontext->uc_mcontext.regs[18] = regs->x18;
-    ucontext->uc_mcontext.regs[19] = regs->x19;
-    ucontext->uc_mcontext.regs[20] = regs->x20;
-    ucontext->uc_mcontext.regs[21] = regs->x21;
-    ucontext->uc_mcontext.regs[22] = regs->x22;
-    ucontext->uc_mcontext.regs[23] = regs->x23;
-    ucontext->uc_mcontext.regs[24] = regs->x24;
-    ucontext->uc_mcontext.regs[25] = regs->x25;
-    ucontext->uc_mcontext.regs[26] = regs->x26;
-    ucontext->uc_mcontext.regs[27] = regs->x27;
-    ucontext->uc_mcontext.regs[28] = regs->x28;
-    ucontext->uc_mcontext.regs[29] = regs->x29;
-    ucontext->uc_mcontext.regs[30] = regs->x30;
-    ucontext->uc_mcontext.sp = regs->sp_el0;
-    ucontext->uc_mcontext.pc = regs->pc;
-    ucontext->uc_mcontext.pstate = regs->cpsr;
-    signal_aarch64_save_fpsimd(&ucontext->uc_mcontext, fpu_ctx);
-}
-
-static bool
-signal_aarch64_restore_ptregs(struct pt_regs *regs,
-                              const signal_aarch64_ucontext_t *ucontext,
-                              fpu_context_t *fpu_ctx) {
-    regs->x0 = ucontext->uc_mcontext.regs[0];
-    regs->x1 = ucontext->uc_mcontext.regs[1];
-    regs->x2 = ucontext->uc_mcontext.regs[2];
-    regs->x3 = ucontext->uc_mcontext.regs[3];
-    regs->x4 = ucontext->uc_mcontext.regs[4];
-    regs->x5 = ucontext->uc_mcontext.regs[5];
-    regs->x6 = ucontext->uc_mcontext.regs[6];
-    regs->x7 = ucontext->uc_mcontext.regs[7];
-    regs->x8 = ucontext->uc_mcontext.regs[8];
-    regs->x9 = ucontext->uc_mcontext.regs[9];
-    regs->x10 = ucontext->uc_mcontext.regs[10];
-    regs->x11 = ucontext->uc_mcontext.regs[11];
-    regs->x12 = ucontext->uc_mcontext.regs[12];
-    regs->x13 = ucontext->uc_mcontext.regs[13];
-    regs->x14 = ucontext->uc_mcontext.regs[14];
-    regs->x15 = ucontext->uc_mcontext.regs[15];
-    regs->x16 = ucontext->uc_mcontext.regs[16];
-    regs->x17 = ucontext->uc_mcontext.regs[17];
-    regs->x18 = ucontext->uc_mcontext.regs[18];
-    regs->x19 = ucontext->uc_mcontext.regs[19];
-    regs->x20 = ucontext->uc_mcontext.regs[20];
-    regs->x21 = ucontext->uc_mcontext.regs[21];
-    regs->x22 = ucontext->uc_mcontext.regs[22];
-    regs->x23 = ucontext->uc_mcontext.regs[23];
-    regs->x24 = ucontext->uc_mcontext.regs[24];
-    regs->x25 = ucontext->uc_mcontext.regs[25];
-    regs->x26 = ucontext->uc_mcontext.regs[26];
-    regs->x27 = ucontext->uc_mcontext.regs[27];
-    regs->x28 = ucontext->uc_mcontext.regs[28];
-    regs->x29 = ucontext->uc_mcontext.regs[29];
-    regs->x30 = ucontext->uc_mcontext.regs[30];
-    regs->sp_el0 = ucontext->uc_mcontext.sp;
-    regs->pc = ucontext->uc_mcontext.pc;
-    regs->cpsr = ucontext->uc_mcontext.pstate;
-    regs->origin_x0 = regs->x0;
-    regs->syscallno = NO_SYSCALL;
-
-    if ((regs->cpsr & 0xF) != 0)
-        regs->cpsr &= ~0xFULL;
-
-    return signal_aarch64_restore_fpsimd(&ucontext->uc_mcontext, fpu_ctx);
-}
-
-static bool signal_aarch64_setup_frame(task_t *task, struct pt_regs *regs,
-                                       int sig, const sigaction_t *action,
-                                       const siginfo_t *info,
-                                       sigset_t restore_mask) {
-    struct pt_regs saved = *regs;
-    signal_aarch64_prepare_syscall_result(&saved, action);
-
-    stack_t frame_altstack;
-    signal_altstack_format_old(&frame_altstack, &task->signal->altstack,
-                               saved.sp_el0);
-
-    uint64_t stack_top = saved.sp_el0;
-    uint64_t trampoline_bytes = 0;
-    bool use_kernel_restorer =
-        !(action->sa_flags & SA_RESTORER) || !action->sa_restorer;
-    if (use_kernel_restorer && !signal_ensure_user_trampoline(task))
-        return false;
-    bool use_altstack =
-        (action->sa_flags & SA_ONSTACK) &&
-        signal_altstack_config_enabled(&task->signal->altstack) &&
-        !signal_altstack_contains_sp(&task->signal->altstack, saved.sp_el0);
-    if (use_altstack) {
-        stack_top = signal_stack_base(&task->signal->altstack) +
-                    task->signal->altstack.ss_size;
-        if (task->signal->altstack.ss_flags & SS_AUTODISARM)
-            signal_altstack_disable(&task->signal->altstack);
-    }
-
-    if (stack_top <= sizeof(signal_aarch64_frame_t) +
-                         sizeof(signal_aarch64_frame_record_t) +
-                         trampoline_bytes)
-        return false;
-
-    uint64_t frame_record_addr = signal_aarch64_align_down(
-        stack_top - sizeof(signal_aarch64_frame_record_t), 16);
-    if (!frame_record_addr)
-        return false;
-
-    uint64_t frame_addr = signal_aarch64_align_down(
-        frame_record_addr - sizeof(signal_aarch64_frame_t), 16);
-    if (!frame_addr || frame_addr >= frame_record_addr)
-        return false;
-
-    signal_aarch64_frame_t frame;
-    memset(&frame, 0, sizeof(frame));
-    memcpy(&frame.info, info, sizeof(frame.info));
-    aarch64_fpu_save(task->arch_context->fpu_ctx);
-    signal_aarch64_fill_ucontext(&frame.ucontext, &saved, restore_mask,
-                                 &frame_altstack, info,
-                                 task->arch_context->fpu_ctx);
-
-    signal_aarch64_frame_record_t frame_record = {
-        .fp = saved.x29,
-        .lr = saved.x30,
-    };
-
-    if (copy_to_user((void *)frame_addr, &frame, sizeof(frame)) ||
-        copy_to_user((void *)frame_record_addr, &frame_record,
-                     sizeof(frame_record))) {
-        return false;
-    }
-
-    memset(regs, 0, sizeof(*regs));
-    regs->pc = (uint64_t)action->sa_handler;
-    regs->sp_el0 = frame_addr;
-    regs->origin_x0 = saved.origin_x0;
-    regs->syscallno = NO_SYSCALL;
-    regs->x0 = sig;
-    if (action->sa_flags & SA_SIGINFO) {
-        regs->x1 = frame_addr + offsetof(signal_aarch64_frame_t, info);
-        regs->x2 = frame_addr + offsetof(signal_aarch64_frame_t, ucontext);
-    }
-    regs->x29 = frame_record_addr;
-    regs->x30 = use_kernel_restorer ? USER_SIGNAL_TRAMPOLINE_START
-                                    : (uint64_t)action->sa_restorer;
-    regs->cpsr = 0x80000300;
-
-    return true;
-}
-#endif
-
-#if defined(__riscv__)
-typedef struct signal_riscv64_user_sigset {
-    uint64_t sig[1];
-} signal_riscv64_user_sigset_t;
-
-typedef struct signal_riscv64_user_regs {
-    uint64_t pc;
-    uint64_t ra;
-    uint64_t sp;
-    uint64_t gp;
-    uint64_t tp;
-    uint64_t t0;
-    uint64_t t1;
-    uint64_t t2;
-    uint64_t s0;
-    uint64_t s1;
-    uint64_t a0;
-    uint64_t a1;
-    uint64_t a2;
-    uint64_t a3;
-    uint64_t a4;
-    uint64_t a5;
-    uint64_t a6;
-    uint64_t a7;
-    uint64_t s2;
-    uint64_t s3;
-    uint64_t s4;
-    uint64_t s5;
-    uint64_t s6;
-    uint64_t s7;
-    uint64_t s8;
-    uint64_t s9;
-    uint64_t s10;
-    uint64_t s11;
-    uint64_t t3;
-    uint64_t t4;
-    uint64_t t5;
-    uint64_t t6;
-} signal_riscv64_user_regs_t;
-
-typedef struct signal_riscv64_ctx_header {
-    uint32_t magic;
-    uint32_t size;
-} signal_riscv64_ctx_header_t;
-
-typedef struct signal_riscv64_f_ext_state {
-    uint32_t f[32];
-    uint32_t fcsr;
-} signal_riscv64_f_ext_state_t;
-
-typedef struct signal_riscv64_d_ext_state {
-    uint64_t f[32];
-    uint32_t fcsr;
-} signal_riscv64_d_ext_state_t;
-
-typedef struct signal_riscv64_extra_ext_header {
-    uint32_t __padding[129] __attribute__((aligned(16)));
-    uint32_t reserved;
-    signal_riscv64_ctx_header_t hdr;
-} signal_riscv64_extra_ext_header_t;
-
-typedef union signal_riscv64_fp_state {
-    signal_riscv64_f_ext_state_t f;
-    signal_riscv64_d_ext_state_t d;
-    signal_riscv64_extra_ext_header_t q;
-} signal_riscv64_fp_state_t;
-
-typedef struct signal_riscv64_sigcontext {
-    signal_riscv64_user_regs_t sc_regs;
-    union {
-        signal_riscv64_fp_state_t sc_fpregs;
-        signal_riscv64_extra_ext_header_t sc_extdesc;
-    };
-} signal_riscv64_sigcontext_t;
-
-typedef struct signal_riscv64_ucontext {
-    uint64_t uc_flags;
-    struct signal_riscv64_ucontext *uc_link;
-    stack_t uc_stack;
-    signal_riscv64_user_sigset_t uc_sigmask;
-    uint8_t __unused[1024 / 8 - sizeof(signal_riscv64_user_sigset_t)];
-    signal_riscv64_sigcontext_t uc_mcontext;
-} signal_riscv64_ucontext_t;
-
-typedef struct signal_riscv64_frame {
-    siginfo_t info;
-    signal_riscv64_ucontext_t ucontext;
-    uint64_t flags;
-    uint64_t syscall_pc;
-} signal_riscv64_frame_t;
-
-_Static_assert(sizeof(signal_riscv64_user_sigset_t) +
-                       sizeof(((signal_riscv64_ucontext_t *)0)->__unused) ==
-                   1024 / 8,
-               "RISC-V user sigset in ucontext must be 1024 bits");
-_Static_assert(offsetof(signal_riscv64_ucontext_t, uc_mcontext) == 176,
-               "RISC-V ucontext.uc_mcontext offset must match the Linux ABI");
-_Static_assert(offsetof(signal_riscv64_sigcontext_t, sc_regs.pc) == 0,
-               "RISC-V mcontext PC must be gregs[0] for musl MC_PC");
-
-static inline uint64_t signal_riscv64_align_down(uint64_t value,
-                                                 uint64_t align) {
-    return value & ~(align - 1);
-}
-
-static inline bool
-signal_riscv64_prepare_syscall_result(struct pt_regs *saved,
-                                      const sigaction_t *action) {
-    if (!saved || !action || saved->syscallno == NO_SYSCALL ||
-        saved->syscallno == SYS_RT_SIGRETURN)
-        return false;
-
-    int64_t retval = (int64_t)saved->a0;
-    if (retval < 0) {
-        switch (-retval) {
-        case ERESTARTNOHAND:
-        case ERESTART_RESTARTBLOCK:
-        case ERESTARTSYS:
-        case ERESTARTNOINTR:
-            saved->a0 = (uint64_t)-EINTR;
-            break;
-        default:
-            break;
-        }
-    }
-
-    return true;
-}
-
-static inline void
-signal_riscv64_fill_ucontext(signal_riscv64_ucontext_t *ucontext,
-                             const struct pt_regs *regs, sigset_t blocked_mask,
-                             const stack_t *altstack,
-                             const fpu_context_t *fpu_ctx) {
-    memset(ucontext, 0, sizeof(*ucontext));
-
-    if (altstack)
-        ucontext->uc_stack = *altstack;
-    ucontext->uc_sigmask.sig[0] = sigset_kernel_to_user(blocked_mask);
-
-    ucontext->uc_mcontext.sc_regs.pc = regs->sepc;
-    ucontext->uc_mcontext.sc_regs.ra = regs->ra;
-    ucontext->uc_mcontext.sc_regs.sp = regs->sp;
-    ucontext->uc_mcontext.sc_regs.gp = regs->gp;
-    ucontext->uc_mcontext.sc_regs.tp = regs->tp;
-    ucontext->uc_mcontext.sc_regs.t0 = regs->t0;
-    ucontext->uc_mcontext.sc_regs.t1 = regs->t1;
-    ucontext->uc_mcontext.sc_regs.t2 = regs->t2;
-    ucontext->uc_mcontext.sc_regs.s0 = regs->s0;
-    ucontext->uc_mcontext.sc_regs.s1 = regs->s1;
-    ucontext->uc_mcontext.sc_regs.a0 = regs->a0;
-    ucontext->uc_mcontext.sc_regs.a1 = regs->a1;
-    ucontext->uc_mcontext.sc_regs.a2 = regs->a2;
-    ucontext->uc_mcontext.sc_regs.a3 = regs->a3;
-    ucontext->uc_mcontext.sc_regs.a4 = regs->a4;
-    ucontext->uc_mcontext.sc_regs.a5 = regs->a5;
-    ucontext->uc_mcontext.sc_regs.a6 = regs->a6;
-    ucontext->uc_mcontext.sc_regs.a7 = regs->a7;
-    ucontext->uc_mcontext.sc_regs.s2 = regs->s2;
-    ucontext->uc_mcontext.sc_regs.s3 = regs->s3;
-    ucontext->uc_mcontext.sc_regs.s4 = regs->s4;
-    ucontext->uc_mcontext.sc_regs.s5 = regs->s5;
-    ucontext->uc_mcontext.sc_regs.s6 = regs->s6;
-    ucontext->uc_mcontext.sc_regs.s7 = regs->s7;
-    ucontext->uc_mcontext.sc_regs.s8 = regs->s8;
-    ucontext->uc_mcontext.sc_regs.s9 = regs->s9;
-    ucontext->uc_mcontext.sc_regs.s10 = regs->s10;
-    ucontext->uc_mcontext.sc_regs.s11 = regs->s11;
-    ucontext->uc_mcontext.sc_regs.t3 = regs->t3;
-    ucontext->uc_mcontext.sc_regs.t4 = regs->t4;
-    ucontext->uc_mcontext.sc_regs.t5 = regs->t5;
-    ucontext->uc_mcontext.sc_regs.t6 = regs->t6;
-
-    memset(&ucontext->uc_mcontext.sc_fpregs, 0,
-           sizeof(ucontext->uc_mcontext.sc_fpregs));
-    if (fpu_ctx) {
-        memcpy(ucontext->uc_mcontext.sc_fpregs.d.f, fpu_ctx->f,
-               sizeof(fpu_ctx->f));
-        ucontext->uc_mcontext.sc_fpregs.d.fcsr = fpu_ctx->fcsr;
-    }
-}
-
-static inline void
-signal_riscv64_restore_ptregs(struct pt_regs *regs,
-                              const signal_riscv64_ucontext_t *ucontext) {
-    regs->ra = ucontext->uc_mcontext.sc_regs.ra;
-    regs->sp = ucontext->uc_mcontext.sc_regs.sp;
-    regs->gp = ucontext->uc_mcontext.sc_regs.gp;
-    regs->tp = ucontext->uc_mcontext.sc_regs.tp;
-    regs->t0 = ucontext->uc_mcontext.sc_regs.t0;
-    regs->t1 = ucontext->uc_mcontext.sc_regs.t1;
-    regs->t2 = ucontext->uc_mcontext.sc_regs.t2;
-    regs->s0 = ucontext->uc_mcontext.sc_regs.s0;
-    regs->s1 = ucontext->uc_mcontext.sc_regs.s1;
-    regs->a0 = ucontext->uc_mcontext.sc_regs.a0;
-    regs->a1 = ucontext->uc_mcontext.sc_regs.a1;
-    regs->a2 = ucontext->uc_mcontext.sc_regs.a2;
-    regs->a3 = ucontext->uc_mcontext.sc_regs.a3;
-    regs->a4 = ucontext->uc_mcontext.sc_regs.a4;
-    regs->a5 = ucontext->uc_mcontext.sc_regs.a5;
-    regs->a6 = ucontext->uc_mcontext.sc_regs.a6;
-    regs->a7 = ucontext->uc_mcontext.sc_regs.a7;
-    regs->s2 = ucontext->uc_mcontext.sc_regs.s2;
-    regs->s3 = ucontext->uc_mcontext.sc_regs.s3;
-    regs->s4 = ucontext->uc_mcontext.sc_regs.s4;
-    regs->s5 = ucontext->uc_mcontext.sc_regs.s5;
-    regs->s6 = ucontext->uc_mcontext.sc_regs.s6;
-    regs->s7 = ucontext->uc_mcontext.sc_regs.s7;
-    regs->s8 = ucontext->uc_mcontext.sc_regs.s8;
-    regs->s9 = ucontext->uc_mcontext.sc_regs.s9;
-    regs->s10 = ucontext->uc_mcontext.sc_regs.s10;
-    regs->s11 = ucontext->uc_mcontext.sc_regs.s11;
-    regs->t3 = ucontext->uc_mcontext.sc_regs.t3;
-    regs->t4 = ucontext->uc_mcontext.sc_regs.t4;
-    regs->t5 = ucontext->uc_mcontext.sc_regs.t5;
-    regs->t6 = ucontext->uc_mcontext.sc_regs.t6;
-    regs->sepc = ucontext->uc_mcontext.sc_regs.pc;
-    regs->sstatus = SIGNAL_RISCV64_USER_SSTATUS;
-    regs->syscallno = NO_SYSCALL;
-}
-
-static bool signal_riscv64_setup_frame(task_t *task, struct pt_regs *regs,
-                                       int sig, const sigaction_t *action,
-                                       const siginfo_t *info,
-                                       sigset_t restore_mask) {
-    struct pt_regs saved = *regs;
-    bool from_syscall = signal_riscv64_prepare_syscall_result(&saved, action);
-
-    stack_t frame_altstack;
-    signal_altstack_format_old(&frame_altstack, &task->signal->altstack,
-                               saved.sp);
-
-    uint64_t stack_top = saved.sp;
-    uint64_t trampoline_bytes = 0;
-    bool use_kernel_restorer =
-        !(action->sa_flags & SA_RESTORER) || !action->sa_restorer;
-    if (use_kernel_restorer && !signal_ensure_user_trampoline(task))
-        return false;
-    bool use_altstack =
-        (action->sa_flags & SA_ONSTACK) &&
-        signal_altstack_config_enabled(&task->signal->altstack) &&
-        !signal_altstack_contains_sp(&task->signal->altstack, saved.sp);
-    if (use_altstack) {
-        stack_top = signal_stack_base(&task->signal->altstack) +
-                    task->signal->altstack.ss_size;
-        if (task->signal->altstack.ss_flags & SS_AUTODISARM)
-            signal_altstack_disable(&task->signal->altstack);
-    }
-
-    if (stack_top <= sizeof(signal_riscv64_frame_t) + trampoline_bytes)
-        return false;
-
-    uint64_t frame_addr = signal_riscv64_align_down(
-        stack_top - sizeof(signal_riscv64_frame_t), 16);
-    if (!frame_addr || frame_addr >= stack_top)
-        return false;
-
-    signal_riscv64_frame_t frame;
-    memset(&frame, 0, sizeof(frame));
-    memcpy(&frame.info, info, sizeof(frame.info));
-    if (from_syscall) {
-        frame.flags |= SIGNAL_RISCV64_FRAME_FROM_SYSCALL;
-        frame.syscall_pc = saved.sepc;
-    }
-    riscv64_fpu_save(task->arch_context->fpu_ctx);
-    signal_riscv64_fill_ucontext(&frame.ucontext, &saved, restore_mask,
-                                 &frame_altstack, task->arch_context->fpu_ctx);
-
-    if (copy_to_user((void *)frame_addr, &frame, sizeof(frame))) {
-        return false;
-    }
-
-    *regs = saved;
-    regs->sepc = (uint64_t)action->sa_handler;
-    regs->ra = use_kernel_restorer ? USER_SIGNAL_TRAMPOLINE_START
-                                   : (uint64_t)action->sa_restorer;
-    regs->sp = frame_addr;
-    regs->a0 = sig;
-    regs->a1 = 0;
-    regs->a2 = 0;
-    if (action->sa_flags & SA_SIGINFO) {
-        regs->a1 = frame_addr + offsetof(signal_riscv64_frame_t, info);
-        regs->a2 = frame_addr + offsetof(signal_riscv64_frame_t, ucontext);
-    }
-    regs->sstatus = (saved.sstatus & ~(1ULL << 8)) | (1ULL << 5);
-    regs->syscallno = NO_SYSCALL;
-
-    return true;
-}
-#endif
+#include <arch/signal_impl.h>
 
 void signal_init() {
     for (int i = 0; i < MAXSIG; i++) {
@@ -1674,13 +586,12 @@ uint64_t sys_sigaction(int sig, const void *action, void *oldaction,
         memset(&new_action, 0, sizeof(new_action));
         new_action.sa_handler = user_action.handler;
         new_action.sa_flags = (int)user_action.flags;
-#if defined(__riscv__) || defined(__loongarch64__)
         new_action.sa_restorer = NULL;
-#else
+#if SIGNAL_ARCH_HAS_RESTORER_FIELD
         new_action.sa_restorer = user_action.restorer;
 #endif
         new_action.sa_mask = sigset_user_to_kernel(user_action.mask);
-#if defined(__x86_64__) || defined(__aarch64__)
+#if SIGNAL_ARCH_VALIDATE_RESTORER
         if (new_action.sa_handler != SIG_DFL &&
             new_action.sa_handler != SIG_IGN &&
             (new_action.sa_flags & SA_RESTORER) != 0 &&
@@ -1708,7 +619,7 @@ uint64_t sys_sigaction(int sig, const void *action, void *oldaction,
             .handler = old_local.sa_handler,
             .flags = (unsigned long)old_local.sa_flags,
             .mask = sigset_kernel_to_user(old_local.sa_mask),
-#if !defined(__riscv__) && !defined(__loongarch64__)
+#if SIGNAL_ARCH_HAS_RESTORER_FIELD
             .restorer = old_local.sa_restorer,
 #endif
         };
@@ -1722,142 +633,7 @@ uint64_t sys_sigaction(int sig, const void *action, void *oldaction,
 }
 
 uint64_t sys_sigreturn(struct pt_regs *regs) {
-#if defined(__x86_64__)
-    arch_disable_interrupt();
-
-    task_t *self = current_task;
-    if (!self || !regs) {
-        return (uint64_t)-EFAULT;
-    }
-
-    ucontext_t frame_ucontext;
-    if (copy_from_user(&frame_ucontext, (void *)regs->rsp,
-                       sizeof(frame_ucontext))) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    if (frame_ucontext.uc_mcontext.fpregs && ({
-            x64_fpu_state_init(self->arch_context->fpu_ctx);
-            copy_from_user(self->arch_context->fpu_ctx,
-                           frame_ucontext.uc_mcontext.fpregs,
-                           signal_x64_restore_fpstate_bytes(&frame_ucontext));
-        })) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-    x64_fpu_restore(self->arch_context->fpu_ctx);
-
-    stack_t restore_altstack = frame_ucontext.uc_stack;
-    restore_altstack.ss_flags &= ~SS_ONSTACK;
-    if (signal_altstack_validate_new(&restore_altstack) < 0) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    signal_x64_restore_ptregs(regs, &frame_ucontext);
-
-    spin_lock(&self->signal->sighand->siglock);
-    self->signal->blocked =
-        sigset_user_to_kernel(frame_ucontext.uc_sigmask.__bits[0]);
-    signal_altstack_store(&self->signal->altstack, &restore_altstack);
-    spin_unlock(&self->signal->sighand->siglock);
-
-    regs->rflags |= (1ULL << 9);
-    regs->r11 = regs->rflags;
-    regs->rcx = regs->rip;
-
-    return regs->rax;
-#elif defined(__aarch64__)
-    arch_disable_interrupt();
-
-    task_t *self = current_task;
-    if (!self || !regs) {
-        return (uint64_t)-EFAULT;
-    }
-
-    if ((regs->sp_el0 & 0xFUL) != 0) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    signal_aarch64_frame_t frame;
-    if (copy_from_user(&frame, (void *)regs->sp_el0, sizeof(frame))) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    stack_t restore_altstack = frame.ucontext.uc_stack;
-    restore_altstack.ss_flags &= ~SS_ONSTACK;
-    if (signal_altstack_validate_new(&restore_altstack) < 0) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    if (!signal_aarch64_restore_ptregs(regs, &frame.ucontext,
-                                       self->arch_context->fpu_ctx)) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-    aarch64_fpu_restore(self->arch_context->fpu_ctx);
-
-    spin_lock(&self->signal->sighand->siglock);
-    self->signal->blocked =
-        sigset_user_to_kernel(frame.ucontext.uc_sigmask.__bits[0]);
-    signal_altstack_store(&self->signal->altstack, &restore_altstack);
-    spin_unlock(&self->signal->sighand->siglock);
-
-    return regs->x0;
-#elif defined(__riscv__)
-    arch_disable_interrupt();
-
-    task_t *self = current_task;
-    if (!self || !regs) {
-        return (uint64_t)-EFAULT;
-    }
-
-    if ((regs->sp & 0xFUL) != 0) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    signal_riscv64_frame_t frame;
-    if (copy_from_user(&frame, (void *)regs->sp, sizeof(frame))) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    stack_t restore_altstack = frame.ucontext.uc_stack;
-    restore_altstack.ss_flags &= ~SS_ONSTACK;
-    if (signal_altstack_validate_new(&restore_altstack) < 0) {
-        task_exit(128 + SIGSEGV);
-        return 0;
-    }
-
-    signal_riscv64_restore_ptregs(regs, &frame.ucontext);
-    if ((frame.flags & SIGNAL_RISCV64_FRAME_FROM_SYSCALL) != 0 &&
-        regs->sepc == frame.syscall_pc) {
-        regs->sepc += SIGNAL_RISCV64_SYSCALL_INS_LEN;
-    }
-    memcpy(self->arch_context->fpu_ctx->f,
-           frame.ucontext.uc_mcontext.sc_fpregs.d.f,
-           sizeof(self->arch_context->fpu_ctx->f));
-    self->arch_context->fpu_ctx->fcsr =
-        frame.ucontext.uc_mcontext.sc_fpregs.d.fcsr;
-    self->arch_context->fpu_ctx->reserved = 0;
-    riscv64_fpu_restore(self->arch_context->fpu_ctx);
-
-    spin_lock(&self->signal->sighand->siglock);
-    self->signal->blocked =
-        sigset_user_to_kernel(frame.ucontext.uc_sigmask.sig[0]);
-    signal_altstack_store(&self->signal->altstack, &restore_altstack);
-    spin_unlock(&self->signal->sighand->siglock);
-
-    return regs->a0;
-#else
-    (void)regs;
-    return (uint64_t)-ENOSYS;
-#endif
+    return signal_arch_sigreturn(regs);
 }
 
 uint64_t sys_sigaltstack(const stack_t *uss, stack_t *uoss) {
@@ -2180,32 +956,12 @@ __attribute__((used)) void task_signal(struct pt_regs *regs) {
             }
         }
 
-#if defined(__x86_64__)
-        if (!signal_x64_setup_frame(self, regs, sig, &action, &info,
-                                    restore_mask)) {
+        if (!signal_arch_setup_frame(self, regs, sig, &action, &info,
+                                     restore_mask)) {
             spin_unlock(&self->signal->sighand->siglock);
             task_exit(128 + SIGSEGV);
             return;
         }
-#elif defined(__aarch64__)
-        if (!signal_aarch64_setup_frame(self, regs, sig, &action, &info,
-                                        restore_mask)) {
-            spin_unlock(&self->signal->sighand->siglock);
-            task_exit(128 + SIGSEGV);
-            return;
-        }
-#elif defined(__riscv__)
-        if (!signal_riscv64_setup_frame(self, regs, sig, &action, &info,
-                                        restore_mask)) {
-            spin_unlock(&self->signal->sighand->siglock);
-            task_exit(128 + SIGSEGV);
-            return;
-        }
-#else
-        spin_unlock(&self->signal->sighand->siglock);
-        task_exit(128 + SIGSYS);
-        return;
-#endif
 
         if (action.sa_flags & SA_RESETHAND) {
             self->signal->sighand->actions[sig].sa_handler = SIG_DFL;
